+++ lib/screens/add_watering/watering_form_notifier.dart
@@ -2,6 +2,8 @@ import 'package:drift/drift.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:plant_application/database/converters.dart';
 import 'package:plant_application/database/plant_app_db.dart';
+import 'package:plant_application/models/fertilizer_data.dart';
+import 'package:plant_application/models/timing_enum.dart';
 import 'package:plant_application/providers/db_provider.dart';
 import 'package:plant_application/providers/plant_provider.dart';
 import 'package:plant_application/screens/add_watering/watering_form_data.dart';
@@ -64,7 +66,7 @@ class WateringFormNotifier extends StateNotifier<WateringFormData> {
     state = state.copyWith(fertilizers: {...state.fertilizers});
   }
 
-  void updatePotSize(int potSize) {
+  void updatePotSize(double potSize) {
     state = state.copyWith(potSize: potSize);
   }
